update problem solving

+++ 05_problem_solving/problem_1.py
@@ -1,25 +1,27 @@
 #############################
 # Problem 1
 # * Problem definition
-# Write a Python program that takes two integers as input and returns their sum.
+# Write a Python program that takes two numbers as input and returns their sum.
 #############################
 # * Problem Instructions
-# Instructions
-# 1- Read the first number → store it in variable num1.
-# 2- Read the second number → store it in variable num2.
-# 3- Calculate sum of 2 numbers in third variable called sum_result
-# 4- Display result on screen
+# 1- Read first number => store in variable num1
+# 2- Read second number => store in variable num2
+# 3- Calculate sum 2 numbers in third variable sum_result
+# 4- display result
 
-# * Problem solution
-num1= int(input("Enter first number: "))
-num2= int(input("Enter second number: "))
+# * Problem Solution
+num1 = float(input("Enter first number: "))
+num2 = float(input("Enter second number: "))
 
 sum_result = num1 + num2
+
 print(f"The sum of {num1} + {num2} = {sum_result}")
 
 
-# * Expected Output
-# The sum of 10 + 5 = 15
+
+
+
+
